otimização dos botões administrador e redirecionamento para pagina de edição

+++ assets/scripts/controller/deleteProduct.js
@@ -6,16 +6,19 @@ managerButton.addEventListener('click', () => {
     const deleteButtons = document.querySelectorAll('.item__button-delete')
     deleteButtons.forEach(button => {
         button.addEventListener('click', async() => {
+            const productId = button.parentElement.id
+            
+            if(productId == ''){
+                alert('Este produto não pode ser excluido pois é um produto modelo.')
+                return
+            }
+
             try {
-                const productId = button.parentElement.getAttribute('id')
                 await productService.deleteProduct(productId)
                 location.reload()
             } catch (error) {
                 console.log(error);
-                alert(`
-                Não foi possível deletar o produto.
-                Verifique se não está tentando deletar um produto modelo "Produto XYZ".
-                `)
+                alert('Não foi possível deletar o produto.')
             }
         })
     })
